Require Jenkins 2.426.3 or newer
https://stats.jenkins.io/pluginversions/badge.html shows that 41% of
installs of 1.9.1 (released 2 years ago) are already running Jenkins
2.426.3 or newer.

https://www.jenkins.io/security/advisory/2024-01-24/#SECURITY-3314
affects Jenkins 2.426.2 and earlier and strongly advises users to upgrade
to 2.426.3.

https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/
recommends either 2.414.3 or 2.426.3 as the Jenkins minimum version.
Reducing the minimum version to 2.414.3 increases the installation
percentage from 41% to 50%, but does not help to persuade users that
they should upgrade to at least 2.426.3.