[JENKINS-73187] - Do not apply icon size to text badges
https://issues.jenkins.io/browse/JENKINS-73187 describes the mistake
that I made when setting the icon class to always include "icon-sm".
Short text badges do not need to be limited by that class because it
causes the text field to be too small for any text to fit.

Fixes #152 as well.

Include a test for the icon class of short text.